HOTFIX
/node delete_instance did not work if ModManager was not in use

+++ plugins/admin/commands.py
@@ -1081,7 +1081,7 @@ Please make sure you forward the following ports:
 
         # uninstall mods
         mod_manager: ModManagerService = ServiceRegistry.get(ModManagerService)
-        if mod_manager.is_running() and instance.server:
+        if mod_manager and mod_manager.is_running() and instance.server:
             for folder, package, version in await get_installed_mods(mod_manager, server=instance.server):
                 await mod_manager.uninstall_package(instance.server, folder, package, version)
